requirements.txt
duckdb>=1.0.0
pandas>=2.2
numpy>=1.26
pyarrow>=16.0
matplotlib>=3.8
pyyaml>=6.0